sls_flask/requirements.txt
flask
raven
psycopg2-binary
python-dotenv
werkzeug
Flask-SQLAlchemy